refactor: reorganize codebase into bounded contexts with layered structure
Move the products module into a `business-units` bounded context split
into domain/application/infrastructure layers, and promote `common/` to
`shared/` as the cross-context kernel.

- domain: Product entity + IProductRepository
- application: GetActiveProducts, GetProductById, GetProductsByBusinessUnit use cases + ProductsFetchException
- infrastructure: PrismaProductRepository, products controller, response DTOs
- shared: PrismaModule (global) and pagination primitives
- update imports, app.module wiring, and tsconfig accordingly
