+++ src/app.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { PrismaModule } from './infrastructure/prisma/prisma.module';
-import { ProductsModule } from './modules/products/products.module';
+import { PrismaModule } from './shared/infrastructure/prisma/prisma.module';
+import { BusinessUnitsModule } from './modules/business-units/business-units.module';
 import { ConfigModule } from '@nestjs/config';
 
 @Module({
-  imports: [ConfigModule.forRoot({ isGlobal: true }), PrismaModule, ProductsModule],
+  imports: [ConfigModule.forRoot({ isGlobal: true }), PrismaModule, BusinessUnitsModule],
 })
 export class AppModule {}
