+++ src/modules/business-units/application/errors/product-fetch.exception.ts
