+++ src/modules/business-units/application/use-cases/get-products-by-business-unit.use-case.ts
@@ -3,10 +3,10 @@ import {
   type IProductRepository,
   PRODUCT_REPOSITORY,
   ProductFilters,
-} from '../../../domain/repositories/product.repository';
-import { Product } from '../../../domain/entities/product.entity';
-import { ProductsFetchException } from '../../../common/exceptions/product-fetch.exception';
-import { CursorPaginatedResult, buildCursorMeta } from '../../../common/pagination/pagination';
+} from '../../domain/repositories/product.repository';
+import { Product } from '../../domain/entities/product.entity';
+import { ProductsFetchException } from '../errors/product-fetch.exception';
+import { CursorPaginatedResult, buildCursorMeta } from '../../../../shared/pagination/pagination';
 
 export interface GetProductsByBusinessUnitInput {
   businessUnitId: string;
