+++ src/modules/business-units/domain/entities/product.entity.ts
