+++ src/modules/business-units/infrastructure/http/controllers/products.controller.ts
@@ -1,11 +1,11 @@
 import { Controller, DefaultValuePipe, Get, Param, ParseIntPipe, Query } from '@nestjs/common';
-import { GetActiveProductsUseCase } from './use-cases/get-active-products.use-case';
-import { GetProductsByBusinessUnitUseCase } from './use-cases/get-products-by-business-unit.use-case';
-import { GetProductByIdUseCase } from './use-cases/get-product-by-id.use-case';
-import { ProductResponseDto } from './dto/product-response.dto';
-import { PaginatedResponseDto } from '../../common/pagination/paginated-response.dto';
-import { sanitizeLimit, DEFAULT_LIMIT } from '../../common/pagination/pagination';
-import { ProductFilters } from '../../domain/repositories/product.repository';
+import { GetActiveProductsUseCase } from '../../../application/use-cases/get-active-products.use-case';
+import { GetProductsByBusinessUnitUseCase } from '../../../application/use-cases/get-products-by-business-unit.use-case';
+import { GetProductByIdUseCase } from '../../../application/use-cases/get-product-by-id.use-case';
+import { ProductResponseDto } from '../dto/product-response.dto';
+import { PaginatedResponseDto } from '../../../../../shared/pagination/paginated-response.dto';
+import { sanitizeLimit, DEFAULT_LIMIT } from '../../../../../shared/pagination/pagination';
+import { ProductFilters } from '../../../domain/repositories/product.repository';
 
 @Controller('products')
 export class ProductsController {
